+++ src/elements/Avatar/tests/index.js
@@ -1,0 +1,4 @@
+import { generateSnapshots } from 'helpers';
+import Avatar from '../index';
+
+generateSnapshots(Avatar);